fix 5

+++ tasks/belov_e_sobel/stl/src/ops_stl.cpp
@@ -2,6 +2,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef>
 #include <cstdint>
 #include <thread>
 #include <vector>
@@ -34,10 +35,10 @@ bool BelovESobelSTL::RunImpl() {
   auto get_px = [&](int col, int row) -> float {
     int clamped_x = std::clamp(col, 0, width - 1);
     int clamped_y = std::clamp(row, 0, height - 1);
-    return static_cast<float>(input[(static_cast<size_t>(clamped_y) * width) + clamped_x]);
+    return static_cast<float>(input[(static_cast<std::size_t>(clamped_y) * width) + clamped_x]);
   };
 
-  int num_threads = static_cast<int>(ppc::util::GetNumThreads());
+  int num_threads = ppc::util::GetNumThreads();
 
   std::vector<std::thread> threads;
   int rows_per_thread = height / num_threads;
@@ -52,7 +53,7 @@ bool BelovESobelSTL::RunImpl() {
                    (1 * get_px(col - 1, row + 1)) + (2 * get_px(col, row + 1)) + (1 * get_px(col + 1, row + 1));
 
         float magnitude = std::sqrt((gx * gx) + (gy * gy));
-        output[(static_cast<size_t>(row) * width) + col] = static_cast<uint8_t>(std::min(255.0F, magnitude));
+        output[(static_cast<std::size_t>(row) * width) + col] = static_cast<uint8_t>(std::min(255.0F, magnitude));
       }
     }
   };
